requirement.txt
Django==4.1.7
celery==5.2.3
redis==4.4.0
Pillow==9.2.0
django-celery-results==2.4.0
